fix: sitemap xml file generation, blog hero with dynamic content and add one review from google
